back-end/requirements.txt
Django==4.2
djangorestframework==3.14.0
mysqlclient==2.1.1
django-cors-headers==4.7.0